requirements.txt
six>=1.0
PyYAML>=3.0
